backend/src/main/java/fr/univcotedazur/simpletcfs/repositories/OrderRepository.java
package fr.univcotedazur.simpletcfs.repositories;

import fr.univcotedazur.simpletcfs.entities.Customer;
import fr.univcotedazur.simpletcfs.entities.Order;
import fr.univcotedazur.simpletcfs.entities.OrderStatus;
import org.springframework.data.domain.Sort;
import org.springframework.data.jpa.repository.JpaRepository;
import org.springframework.data.jpa.repository.Query;
import org.springframework.data.repository.query.Param;
import org.springframework.stereotype.Repository;

import java.util.List;

@Repository
public interface OrderRepository extends JpaRepository<Order, Long> {

    @Query("SELECT o FROM Order o WHERE o.customer = :customer and o.status = :status")
    public List<Order> findOrdersForCustomerWithStatus(
            @Param("customer") Customer customer,
            @Param("status") OrderStatus state,
            Sort sort);

}
